Test: 테스트코드 - 통합테스트코드 작성
